quickmeal_project/requirements.txt
Babel==2.3.4
Django==1.9.12
django-appconf==1.0.2
django-compressor==2.1
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.3
django-tables2==1.0.7
django-template-debug==0.3.5
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
factory-boy==2.8.1
Faker==0.7.7
funcsigs==1.0.2
ipaddress==1.0.17
mock==2.0.0
pbr==1.10.0
phonenumbers==7.7.5
Pillow==3.4.2
pkg-resources==0.0.0
purl==1.3
pycountry==1.20
python-dateutil==2.6.0
pytz==2016.10
razorpay==0.2.6
rcssmin==1.0.6
requests==2.12.4
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.19
